+++ Tests/CoverageHost/AdiumY/AIAccount.h
@@ -1,0 +1,1 @@
+../../../Frameworks/Adium/Source/AIAccount.h