+++ Tests/CoverageHost/AdiumY/AIContentObject.h
@@ -1,0 +1,1 @@
+../../../Frameworks/Adium/Source/AIContentObject.h